Move all server error related functions to useEventServerErrors hook to make them more reusable.
