+++ src/domain/event/__tests__/utils.test.ts
@@ -1,5 +1,4 @@
 /* eslint-disable max-len */
-import i18n from 'i18next';
 import { advanceTo, clear } from 'jest-date-mock';
 
 import { EXTLINK, WEEK_DAY } from '../../../constants';
@@ -41,7 +40,6 @@ import {
   getEventInfoLanguages,
   getEventInitialValues,
   getEventPayload,
-  getEventServerErrors,
   getEventTimes,
   getNewEventTimes,
   getRecurringEventPayload,
@@ -170,156 +168,6 @@ describe('sortWeekDays function', () => {
       WEEK_DAY.FRI,
       WEEK_DAY.SAT,
       WEEK_DAY.SUN,
-    ]);
-  });
-});
-
-describe('getEventServerErrors', () => {
-  it('should return server error items', () => {
-    expect(
-      getEventServerErrors({
-        eventType: EVENT_TYPE.General,
-        result: {
-          audience_min_age: ['Tämän luvun on oltava vähintään 0.'],
-          audience_max_age: ['Tämän luvun on oltava vähintään 0.'],
-          description: {
-            fi: 'This field must be specified before an event is published.',
-          },
-          end_time: [
-            'End time cannot be in the past. Please set a future end time.',
-          ],
-          external_links: [{ link: ['Syötä oikea URL-osoite.'] }],
-          keywords: [
-            'This field must be specified before an event is published.',
-          ],
-          location: [
-            'This field must be specified before an event is published.',
-          ],
-          maximum_attendee_capacity: ['Tämän luvun on oltava vähintään 0.'],
-          minimum_attendee_capacity: ['Tämän luvun on oltava vähintään 0.'],
-          name: ['The name must be specified.'],
-          offers: [
-            'Price info must be specified before an event is published.',
-          ],
-          short_description: {
-            fi: 'Short description length must be 160 characters or less',
-          },
-          start_time: [
-            'This field must be specified before an event is published.',
-          ],
-          videos: [
-            {
-              name: ['Tämä kenttä ei voi olla tyhjä.'],
-              url: ['Syötä oikea URL-osoite.'],
-            },
-          ],
-          dummy1: {
-            error: ['Not translated error message.'],
-          },
-          dummy2: {
-            error: 'Not translated error message.',
-          },
-        },
-        t: i18n.t.bind(i18n),
-      })
-    ).toEqual([
-      { label: 'Alaikäraja', message: 'Tämän luvun on oltava vähintään 0.' },
-      {
-        label: 'Yläikäraja',
-        message: 'Tämän luvun on oltava vähintään 0.',
-      },
-      {
-        label: 'Tapahtuman kuvaus suomeksi',
-        message: 'Tämä kenttä on määritettävä ennen tapahtuman julkaisemista.',
-      },
-      {
-        label: 'Tapahtuma päättyy',
-        message:
-          'Lopetusaika ei voi olla menneisyydessä. Määritä tuleva päättymisaika.',
-      },
-      {
-        label: 'Tapahtuma sosiaalisessa mediassa',
-        message: 'Syötä oikea URL-osoite.',
-      },
-      {
-        label: 'Avainsanahaku',
-        message: 'Tämä kenttä on määritettävä ennen tapahtuman julkaisemista.',
-      },
-      {
-        label: 'Paikka',
-        message: 'Tämä kenttä on määritettävä ennen tapahtuman julkaisemista.',
-      },
-      {
-        label: 'Enimmäisosallistujamäärä',
-        message: 'Tämän luvun on oltava vähintään 0.',
-      },
-      {
-        label: 'Vähimmäisosallistujamäärä',
-        message: 'Tämän luvun on oltava vähintään 0.',
-      },
-      {
-        label: 'Tapahtuman otsikko ',
-        message: 'Nimi on pakollinen.',
-      },
-      {
-        label: 'Tapahtuman hintatiedot',
-        message: 'Hintatiedot on määritettävä ennen tapahtuman julkaisemista.',
-      },
-      {
-        label: 'Lyhyt kuvaus suomeksi (korkeintaan 160 merkkiä)',
-        message: 'Lyhyen kuvauksen pituus saa olla enintään 160 merkkiä',
-      },
-      {
-        label: 'Tapahtuma alkaa',
-        message: 'Tämä kenttä on määritettävä ennen tapahtuman julkaisemista.',
-      },
-      {
-        label: 'Videon nimi',
-        message: 'Tämä kenttä ei voi olla tyhjä.',
-      },
-      {
-        label: 'Videon URL-osoite',
-        message: 'Syötä oikea URL-osoite.',
-      },
-      {
-        label: 'event.form.labelDummy1',
-        message: 'Not translated error message.',
-      },
-      {
-        label: 'event.form.labelDummy2',
-        message: 'Not translated error message.',
-      },
-    ]);
-  });
-  it('should return server error items when result is array', () => {
-    expect(
-      getEventServerErrors({
-        eventType: EVENT_TYPE.General,
-        result: [
-          { name: ['The name must be specified.'] },
-          { name: ['The name must be specified.'] },
-          { name: ['The name must be specified.'] },
-          { name: ['The name must be specified.'] },
-        ],
-        t: i18n.t.bind(i18n),
-      })
-    ).toEqual([
-      {
-        label: 'Tapahtuman otsikko ',
-        message: 'Nimi on pakollinen.',
-      },
-      {
-        label: 'Tapahtuman otsikko ',
-        message: 'Nimi on pakollinen.',
-      },
-      {
-        label: 'Tapahtuman otsikko ',
-        message: 'Nimi on pakollinen.',
-      },
-      {
-        label: 'Tapahtuman otsikko ',
-        message: 'Nimi on pakollinen.',
-      },
     ]);
   });
 });
