+++ src/domain/event/utils.ts
@@ -56,17 +56,10 @@ import {
   SuperEventType,
   UserFieldsFragment,
 } from '../../generated/graphql';
-import {
-  Language,
-  LEServerError,
-  PathBuilderProps,
-  ServerErrorItem,
-} from '../../types';
+import { Language, PathBuilderProps } from '../../types';
 import getLocalisedString from '../../utils/getLocalisedString';
 import getNextPage from '../../utils/getNextPage';
 import getPathBuilder from '../../utils/getPathBuilder';
-import lowerCaseFirstLetter from '../../utils/lowerCaseFirstLetter';
-import pascalCase from '../../utils/pascalCase';
 import queryBuilder from '../../utils/queryBuilder';
 import sanitizeHtml from '../../utils/sanitizeHtml';
 import {
@@ -517,198 +510,6 @@ export const addImageSchema = Yup.object().shape(
   },
   [[ADD_IMAGE_FIELDS.SELECTED_IMAGE, ADD_IMAGE_FIELDS.URL]]
 );
-
-const getEventServerErrorLabel = ({
-  eventType,
-  key,
-  langText,
-  t,
-}: {
-  eventType: string;
-  key: string;
-  langText?: string;
-  t: TFunction;
-}) => {
-  switch (key) {
-    case 'end_time':
-    case 'name':
-    case 'start_time':
-      return t(`event.form.label${pascalCase(key)}.${eventType}`);
-    case 'description':
-    case 'short_description':
-      return t(`event.form.label${pascalCase(key)}.${eventType}`, { langText });
-    case 'external_links':
-      return t(`event.form.titleSocialMedia.${eventType}`);
-    case 'offers':
-      return t(`event.form.titlePriceInfo.${eventType}`);
-    default:
-      return t(`event.form.label${pascalCase(key)}`);
-  }
-};
-
-const getEventServerErrorMessage = ({
-  error,
-  t,
-}: {
-  error: LEServerError;
-  t: TFunction;
-}) => {
-  let errorStr = '';
-  if (Array.isArray(error)) {
-    const e =
-      typeof error[0] === 'object'
-        ? Object.values(error[0]).find((item) => item)
-        : error[0];
-    errorStr = Array.isArray(e) ? e[0] : e;
-  } else {
-    const e = Object.values(error).find((item) => item);
-    errorStr = Array.isArray(e) ? e[0] : e;
-  }
-
-  switch (errorStr) {
-    case 'End time cannot be in the past. Please set a future end time.':
-      return t(`event.serverError.endTimeInPast`);
-    case 'Price info must be specified before an event is published.':
-      return t(`event.serverError.offersIsRequired`);
-    case 'Short description length must be 160 characters or less':
-      return t(`event.serverError.shortDescriptionTooLong`);
-    case 'Syötä oikea URL-osoite.':
-      return t(`event.serverError.invalidUrl`);
-    case 'The name must be specified.':
-      return t(`event.serverError.nameIsRequired`);
-    case 'This field must be specified before an event is published.':
-      return t(`event.serverError.requiredWhenPublishing`);
-    case 'Tämä kenttä ei voi olla tyhjä.':
-      return t(`event.serverError.required`);
-    case 'Tämän luvun on oltava vähintään 0.':
-      return t(`event.serverError.min0`);
-    default:
-      return errorStr;
-  }
-};
-
-const getLocalizedServerError = ({
-  error,
-  key,
-  t,
-  eventType,
-}: {
-  error: LEServerError;
-  eventType: string;
-  key: string;
-  t: TFunction;
-}): ServerErrorItem[] => {
-  /* istanbul ignore else */
-  if (typeof error === 'object') {
-    return Object.entries(error).reduce(
-      (previous: ServerErrorItem[], [lang, e]) => [
-        ...previous,
-        {
-          label: getEventServerErrorLabel({
-            key,
-            eventType,
-            langText: lowerCaseFirstLetter(t(`form.inLanguage.${lang}`)),
-            t,
-          }),
-          message: getEventServerErrorMessage({ error: [e] as string[], t }),
-        },
-      ],
-      []
-    );
-  } else {
-    return [];
-  }
-};
-
-const getVideoServerError = ({
-  error,
-  t,
-}: {
-  error: LEServerError;
-  t: TFunction;
-}): ServerErrorItem[] => {
-  /* istanbul ignore else */
-  if (Array.isArray(error)) {
-    return Object.entries(error[0]).reduce(
-      (previous: ServerErrorItem[], [key, e]) => [
-        ...previous,
-        {
-          label: t(`event.form.labelVideo${pascalCase(key)}`),
-          message: getEventServerErrorMessage({ error: e as string[], t }),
-        },
-      ],
-      []
-    );
-  } else {
-    return [];
-  }
-};
-
-const getEventServerError = ({
-  error,
-  key,
-  t,
-  eventType,
-}: {
-  error: LEServerError;
-  eventType: string;
-  key: string;
-  t: TFunction;
-}): ServerErrorItem[] => {
-  switch (key) {
-    case 'description':
-    case 'short_description':
-      return getLocalizedServerError({ error, eventType, key, t });
-    case 'videos':
-      return getVideoServerError({ error, t });
-    default:
-      return [
-        {
-          label: getEventServerErrorLabel({
-            key,
-            eventType,
-            t,
-          }),
-          message: getEventServerErrorMessage({ error, t }),
-        },
-      ];
-  }
-};
-
-export const getEventServerErrors = ({
-  eventType,
-  result,
-  t,
-}: {
-  eventType: string;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  result: Record<string, any> | Record<string, any>[];
-  t: TFunction;
-}): ServerErrorItem[] =>
-  Array.isArray(result)
-    ? result.reduce(
-        (previous: ServerErrorItem[], r) => [
-          ...previous,
-          ...getEventServerErrors({
-            eventType,
-            result: r,
-            t,
-          }),
-        ],
-        []
-      )
-    : Object.entries(result).reduce(
-        (previous: ServerErrorItem[], [key, error]) => [
-          ...previous,
-          ...getEventServerError({
-            error,
-            eventType,
-            key,
-            t,
-          }),
-        ],
-        []
-      );
 
 export const eventPathBuilder = ({
   args,
